pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ananke"
version = "0.1.0"
description = "A comprehensive timeseries experimentation framework"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Collective AI", email = "info@collective.ai"},
]
keywords = ["timeseries", "machine-learning", "forecasting", "benchmarking"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Core data science
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "polars>=0.20.0",
    "scikit-learn>=1.3.0",
    # Deep learning
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    # Timeseries specific
    "darts>=0.27.0",
    "prophet>=1.1.0",
    "catboost>=1.2.0",
    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    # Configuration and validation
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "hydra-core>=1.3.0",
    # Experiment tracking
    "mlflow>=2.8.0",
    # Workflow orchestration
    "dagster>=1.5.0",
    # Utilities
    "tqdm>=4.65.0",
    "more-itertools>=10.0.0",
    "typer>=0.9.0",
    "dagster-celery>=0.27.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.2.0",
    "mkdocstrings[python]>=0.22.0",
]

all = [
    "ananke[dev,docs]",
]

[project.urls]
Homepage = "https://github.com/collective-ai/ananke"
Documentation = "https://ananke.readthedocs.io"
Repository = "https://github.com/collective-ai/ananke"
Issues = "https://github.com/collective-ai/ananke/issues"

[project.scripts]
ananke = "anankecli:main"

[tool.hatch.build.targets.wheel]
packages = ["ananke"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=ananke",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]
